% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_by_read_ratio_deviation.R
\name{classify_by_read_ratio_deviation}
\alias{classify_by_read_ratio_deviation}
\title{Classify SNPs by read ratio deviation}
\usage{
classify_by_read_ratio_deviation(het, verbose = FALSE)
}
\arguments{
\item{het}{A data.frame including columns \code{Het.allele.deviation} and \code{EM_class}.}

\item{verbose}{Logical, whether to print info.}
}
\value{
A modified \code{het} with updated \code{EM_class} and \code{allele.deviation.seed}.
}
\description{
Uses a normal distribution CI based on single-copy SNPs to reclassify ambiguous SNPs.
}
