% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_LLR.R
\name{classify_snps_by_llr}
\alias{classify_snps_by_llr}
\title{Classify SNPs based on log-likelihood ratio}
\usage{
classify_snps_by_llr(het, probs1, probs2, cutoff = 0.99)
}
\arguments{
\item{het}{A data.frame containing \code{probs1} (K) and \code{probs2} (D) columns.}

\item{probs1}{Vector of likelihoods under the single-copy model.}

\item{probs2}{Vector of likelihoods under the multicopy model.}

\item{cutoff}{Numeric LLR threshold corresponding to 0.99/0.01 (default).}
}
\value{
A modified \code{het} data.frame with \code{L1}, \code{L2}, \code{LLR}, and \code{EM_class} columns.
}
\description{
Classifies each SNP based on the LLR between the two EM models.
}
