% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_initial_model.R
\name{fit_initial_model}
\alias{fit_initial_model}
\title{Fit initial VGAM model}
\usage{
fit_initial_model(
  regData2,
  weight1,
  weight2,
  boundfit = FALSE,
  lboundary = NULL,
  uboundary = NULL,
  verbose = FALSE
)
}
\arguments{
\item{regData2}{Data.frame used for modeling (Z, MAF, het, N).}

\item{weight1}{Vector of weights for class K.}

\item{weight2}{Vector of weights for class D.}

\item{boundfit}{Logical: whether to enforce boundary constraints.}

\item{lboundary}{Lower boundary for slope (if \code{boundfit = TRUE}).}

\item{uboundary}{Upper boundary for slope (if \code{boundfit = TRUE}).}

\item{verbose}{Whether to print debug messages.}
}
\value{
A list with the fitted model (\code{fit3}), coefficient vector (\code{coef_fit3}), and flag (\code{offsetfit}).
}
\description{
Fits a betabinomial model using VGAM on the initial EM input data.
}
