% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_cutoff.R
\name{plot_distance_cutoff}
\alias{plot_distance_cutoff}
\title{Plot SNP distance cutoff and geometric model}
\usage{
plot_distance_cutoff(distances, dist_cutoff, p1, p2, outpath = ".", ID = "")
}
\arguments{
\item{distances}{Vector of pairwise distances.}

\item{dist_cutoff}{Cutoff to highlight on plot.}

\item{p1}{Fitted p1 from EM.}

\item{p2}{Fitted p2 from EM.}

\item{outpath}{Output directory.}

\item{ID}{Optional run ID for filename.}
}
\description{
Visualizes observed and fitted distance distributions between multicopy SNPs.
}
