% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EM_it.R
\name{plot_EM_it}
\alias{plot_EM_it}
\title{Plot EM iteration results}
\usage{
plot_EM_it(het, fit, weight1, rs, iteration, outpath, ID = "")
}
\arguments{
\item{het}{The original het data frame.}

\item{fit}{The VGAM fit object from the EM loop.}

\item{weight1}{A vector of weights for class K.}

\item{rs}{Indices of SNPs used for plotting (subsample).}

\item{iteration}{Integer indicating the EM iteration number.}

\item{outpath}{Path to output folder.}

\item{ID}{Optional sample or run ID to use in the filename.}
}
\value{
A PDF file with two diagnostic plots.
}
\description{
Generates plots of fitted heterozygosity vs. MAF, with model predictions overlaid.
}
