% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_ratio_deviation.R
\name{plot_read_ratio_deviation}
\alias{plot_read_ratio_deviation}
\title{Plot read ratio deviation and allele ratio density}
\usage{
plot_read_ratio_deviation(het, rs, outpath, ID = "", q1 = NULL, q2 = NULL)
}
\arguments{
\item{het}{A data.frame including EM_class, Het.allele.ratio, Het.allele.deviation, Heterozygous.geno.freq.}

\item{rs}{A vector of row indices (e.g. subsample) to plot.}

\item{outpath}{Directory to save output.}

\item{ID}{Label for filenames.}

\item{q1}{Optional lower cutoff line (e.g. from CI).}

\item{q2}{Optional upper cutoff line.}
}
\description{
Visualizes read ratio deviations and allele ratio distributions across EM classes.
}
