% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_em_input.R
\name{prepare_em_input}
\alias{prepare_em_input}
\title{Prepare EM input data}
\usage{
prepare_em_input(het, subsample = 10000, verbose = FALSE)
}
\arguments{
\item{het}{A data.frame with heterozygosity statistics.}

\item{subsample}{Integer: number of variants to subsample for plotting (default = 10000).}

\item{verbose}{Whether to print debug output.}
}
\value{
A list containing regData, regData2, weights, prediction data, and sampling index.
}
\description{
Prepares regression data and weights for the ParaMaskEM algorithm.
}
